# vlife-admin/src/main/resources/static/doc/openapi.json
{
  "components": {
    "schemas": {
...
      "SysUser": {
        "properties": {
          "id": {
            "format": "string",
            "description": "主键id",
            "type": "string"
          },
          "status": {
            "format": "string",
            "description": "实体状态 【1有效 0逻辑删除】",
            "type": "string"
          },
          "createDate": {
            "format": "string",
            "description": "记录创建日期",
            "type": "string"
          },
          "modifyDate": {
            "format": "string",
            "description": "最近修改日期",
            "type": "string"
          },
          "name": {
            "format": "string",
            "description": "姓名",
            "type": "string"
          },
          "username": {
            "format": "string",
            "description": "用户名",
            "type": "string"
          },
          "password": {
            "format": "string",
            "description": "密码",
            "type": "string"
          },
          "sysGroupId": {
            "format": "string",
            "description": "权限组",
            "type": "string"
          },
          "idno": {
            "format": "string",
            "description": "证件号",
            "type": "string"
          },
          "tel": {
            "format": "string",
            "description": "电话",
            "type": "string"
          },
          "usetype": {
            "format": "string",
            "description": "用户类型",
            "type": "string"
          },
          "state": {
            "format": "string",
            "description": "启用状态",
            "type": "string"
          }
        }
      },
...
      "PageVoUserVo": {
        "properties": {
          "result": {
            "description": "No comments found.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserVo"
            }
          },
          "size": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer"
          },
          "page": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number"
          },
          "totalPage": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number"
          },
          "first": {
            "format": "boolean",
            "description": "No comments found.",
            "type": "boolean"
          },
          "last": {
            "format": "boolean",
            "description": "No comments found.",
            "type": "boolean"
          }
        }
      },
...
      "UserVo": {
        "properties": {
          "id": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "name": {
            "format": "string",
            "description": "用户名",
            "type": "string"
          },
          "username": {
            "format": "string",
            "description": "账号",
            "type": "string"
          },
          "groupName": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "idno": {
            "format": "string",
            "description": "证件号码",
            "type": "string"
          },
          "tel": {
            "format": "string",
            "description": "联系电话",
            "type": "string"
          },
          "usetype": {
            "format": "string",
            "description": "用户类型",
            "type": "string"
          },
          "state": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        }
      },
...
      "PageVoSysUser": {
        "properties": {
          "result": {
            "description": "No comments found.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SysUser"
            }
          },
          "size": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer"
          },
          "page": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number"
          },
          "totalPage": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number"
          },
          "first": {
            "format": "boolean",
            "description": "No comments found.",
            "type": "boolean"
          },
          "last": {
            "format": "boolean",
            "description": "No comments found.",
            "type": "boolean"
          }
        }
      },
...
      "UserDetailVo": {
        "properties": {
          "id": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "name": {
            "format": "string",
            "description": "用户名",
            "type": "string"
          },
          "username": {
            "format": "string",
            "description": "账号",
            "type": "string"
          },
          "groupName": {
            "format": "string",
            "description": "角色组名称",
            "type": "string"
          },
          "resourceCodes": {
            "format": "string",
            "description": "用户有的权限资源代码",
            "type": "string"
          },
          "menus": {
            "format": "string",
            "description": "menus",
            "type": "string"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8088"
    }
  ],
  "openapi": "3.0.3",
  "paths": {
...
    "/sysUser/save": {
      "post": {
        "summary": "保存用户表;",
        "deprecated": false,
        "description": "保存用户表;",
        "tags": [
          "用户表接口;"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysUser"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": \"144\",\n  \"status\": \"yfey4t\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"name\": \"ellie.kuphal\",\n  \"username\": \"ellie.kuphal\",\n  \"password\": \"d3zwyj\",\n  \"sysGroupId\": \"144\",\n  \"idno\": \"fa7btc\",\n  \"tel\": \"hzsgoq\",\n  \"usetype\": \"e7hnot\",\n  \"state\": \"tk6309\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SysUser"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"status\": \"tr5t8b\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"name\": \"ellie.kuphal\",\n  \"username\": \"ellie.kuphal\",\n  \"password\": \"qmdg4n\",\n  \"sysGroupId\": \"144\",\n  \"idno\": \"4hff29\",\n  \"tel\": \"k8gydq\",\n  \"usetype\": \"w4j71e\",\n  \"state\": \"f830nx\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "save",
        "parameters": []
      }
    },
...
    "/sysUser/page": {
      "get": {
        "deprecated": false,
        "description": "",
        "tags": [
          "用户表接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageVoSysUser"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"result\": [\n    {\n      \"id\": \"144\",\n      \"status\": \"ikjlc0\",\n      \"createDate\": \"2022-08-13\",\n      \"modifyDate\": \"2022-08-13\",\n      \"name\": \"ellie.kuphal\",\n      \"username\": \"ellie.kuphal\",\n      \"password\": \"qqdyzn\",\n      \"sysGroupId\": \"144\",\n      \"idno\": \"iymviz\",\n      \"tel\": \"smxwe8\",\n      \"usetype\": \"v80ixm\",\n      \"state\": \"talort\"\n    }\n  ],\n  \"size\": 10,\n  \"page\": 1,\n  \"total\": 12,\n  \"totalPage\": 296,\n  \"first\": true,\n  \"last\": true\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "page",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "用户姓名",
            "required": false,
            "example": "fm1cok",
            "name": "search"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "角色组名称",
            "required": false,
            "example": "ellie.kuphal",
            "name": "sysGroup_name"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "example": "hdsga0",
            "name": "state"
          }
        ]
      }
    },
    "/sysUser/detail/{id}": {
      "get": {
        "summary": "明细查询用户表(视图);",
        "deprecated": false,
        "description": "明细查询用户表(视图);",
        "tags": [
          "用户表接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SysUser"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"status\": \"kme14d\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"name\": \"ellie.kuphal\",\n  \"username\": \"ellie.kuphal\",\n  \"password\": \"tbzl08\",\n  \"sysGroupId\": \"144\",\n  \"idno\": \"fchimn\",\n  \"tel\": \"oigh6f\",\n  \"usetype\": \"mdqvhl\",\n  \"state\": \"cbivqw\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "detail",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
    "/sysUser/detail/userDetailVo/{id}": {
      "get": {
        "deprecated": false,
        "description": "",
        "tags": [
          "用户表接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetailVo"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"name\": \"ellie.kuphal\",\n  \"username\": \"ellie.kuphal\",\n  \"groupName\": \"ellie.kuphal\",\n  \"resourceCodes\": [\n    \"w35zkj\"\n  ],\n  \"menus\": [\n    \"w0dpw5\"\n  ]\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "usesrDetailVoDetail",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "No comments found.",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
    "/sysUser/page/userVo": {
      "get": {
        "summary": "分页查询用户表(视图);",
        "deprecated": false,
        "description": "分页查询用户表(视图);",
        "tags": [
          "用户表接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageVoUserVo"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"result\": [\n    {\n      \"id\": \"144\",\n      \"name\": \"ellie.kuphal\",\n      \"username\": \"ellie.kuphal\",\n      \"groupName\": \"ellie.kuphal\",\n      \"idno\": \"wat37w\",\n      \"tel\": \"llkjmz\",\n      \"usetype\": \"o6w2pg\",\n      \"state\": \"nt4852\"\n    }\n  ],\n  \"size\": 10,\n  \"page\": 1,\n  \"total\": 807,\n  \"totalPage\": 211,\n  \"first\": true,\n  \"last\": true\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pageUser",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "用户姓名",
            "required": false,
            "example": "k794i6",
            "name": "search"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "角色组名称",
            "required": false,
            "example": "ellie.kuphal",
            "name": "sysGroup_name"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "example": "oddr2h",
            "name": "state"
          }
        ]
      }
    },
...
    "/sysUser/remove/{id}": {
      "delete": {
        "summary": "逻辑删除;",
        "deprecated": false,
        "description": "逻辑删除;",
        "tags": [
          "用户表接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "974"
                  }
                }
              }
            }
          }
        },
        "operationId": "remove",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "test",
    "version": "1.0.0"
  },
...
}